Added aliases. Re-wrote parser for Wikipedia / UN Population data.
Combined with the format for aliases, the new parser ing method Wikipedia should be super useful to re-use parsing other sites to scrape data.

I'm aware the import script still needs to be broken up.

Removed "Netherlands Antilles" as it was dissolved in 2010 (seems like most the countries it was broken up into are there - added Curacao as it was missing).