Replace 'humanRights' field with 'warnings' field.
As per issue #10 how warnings are raised has changed and the humanRights field has been simplified and consolidated back into the 'warnings' property - which has been reinstated in a refactored form to give it more utility in future.

After a review of the data the threshold for raising a warning has been set fairly conservatively. Warnings are only raised if there is a clear case for doing so.