Much improved importing/updating behaviour.
If countries are edited or removed from countries.csv the DB is up now updated with the changed data without having to be completely reset.

This makes it much easier to correct or update information without impacting the live service or having to refetch information from other sites (this is useful particularly as external sites can be unreliable and it's preferable to be able to persist any info obtained from them).