Added custom curated list of capital cities in CSV.
This is based on values taken from the CIA World Factbook but is more complete, in that it has data for more states (with the exception of uninhabited locations or administrative regions; indicated in the "notes" column).

It only has one capital per country specified, but the intention is for it to be correct and to include multiple capitals where there exist, with details in the notes field.